Refactor package structure and rename main class
