+++ README.md
@@ -1,0 +1,37 @@
+# Shattered Echoes
+
+Welcome to the Shattered Echoes project! This README will guide you through the setup and usage of the project.
+
+## Table of Contents
+- [Introduction](#introduction)
+- [Installation](#installation)
+- [Usage](#usage)
+- [Contributing](#contributing)
+- [License](#license)
+
+## Introduction
+Shattered Echoes is a Java project designed to...
+
+## Installation
+To install the project, follow these steps:
+1. Clone the repository:
+    ```bash
+    git clone https://github.com/yourusername/shattered_echoes.git
+    ```
+2. Navigate to the project directory:
+    ```bash
+    cd shattered_echoes
+    ```
+3. Build the project using your preferred build tool.
+
+## Usage
+To run the project, execute the following command:
+```bash
+java -jar shattered_echoes.jar
+```
+
+## Contributing
+Contributions are welcome! Please fork the repository and submit a pull request.
+
+## License
+This project is licensed under the MIT License. See the [LICENSE](LICENSE) file for details.
