+++ src/main/java/com/Main/App.java
@@ -1,6 +1,4 @@
-
-
-package com.csit227;
+package com.Main;
 
 /**
  * Hello world!
