+++ src/main/java/com/characters/NPC.java
@@ -1,0 +1,5 @@
+package com.characters;
+
+public class NPC {
+    
+}
