+++ src/main/java/com/characters/Player.java
@@ -1,0 +1,5 @@
+package com.characters;
+
+public class Player {
+    
+}
