+++ src/main/java/com/combat/Combat.java
@@ -1,0 +1,5 @@
+package com.combat;
+
+public class Combat {
+    
+}
