+++ src/main/java/com/combat/Enemy.java
@@ -1,0 +1,5 @@
+package com.combat;
+
+public class Enemy {
+    
+}
