+++ src/main/java/com/items/Item.java
@@ -1,0 +1,5 @@
+package com.items;
+
+public class Item {
+    
+}
