+++ src/main/java/com/items/Potion.java
@@ -1,0 +1,5 @@
+package com.items;
+
+public class Potion {
+    
+}
