+++ src/main/java/com/items/Weapon.java
@@ -1,0 +1,5 @@
+package com.items;
+
+public class Weapon {
+    
+}
