+++ src/main/java/com/quests/Quest.java
@@ -1,0 +1,5 @@
+package com.quests;
+
+public class Quest {
+    
+}
