+++ src/main/java/com/quests/QuestManager.java
@@ -1,0 +1,5 @@
+package com.quests;
+
+public class QuestManager {
+    
+}
