+++ src/main/java/com/story/Dialogue.java
@@ -1,0 +1,5 @@
+package com.story;
+
+public class Dialogue {
+    
+}
