+++ src/main/java/com/utils/InputHandler.java
@@ -1,0 +1,5 @@
+package com.utils;
+
+public class InputHandler {
+    
+}
